Google related bounds tied to SearchMapWithGoogleMap implementation
