+++ src/util/googleMaps.js
@@ -90,33 +90,6 @@ export const getPlacePredictions = (search, sessionToken) =>
   });
 
 /**
- * Convert Google formatted LatLng object to Sharetribe SDK's LatLng coordinate format
- *
- * @param {LatLng} googleLatLng - Google Maps LatLng
- *
- * @return {SDKLatLng} - Converted latLng coordinate
- */
-export const googleLatLngToSDKLatLng = googleLatLng => {
-  return new SDKLatLng(googleLatLng.lat(), googleLatLng.lng());
-};
-
-/**
- * Convert Google formatted bounds object to Sharetribe SDK's bounds format
- *
- * @param {LatLngBounds} googleBounds - Google Maps LatLngBounds
- *
- * @return {SDKLatLngBounds} - Converted bounds
- */
-export const googleBoundsToSDKBounds = googleBounds => {
-  if (!googleBounds) {
-    return null;
-  }
-  const ne = googleBounds.getNorthEast();
-  const sw = googleBounds.getSouthWest();
-  return new SDKLatLngBounds(new SDKLatLng(ne.lat(), ne.lng()), new SDKLatLng(sw.lat(), sw.lng()));
-};
-
-/**
  * Cut some precision from bounds coordinates to tackle subtle map movements
  * when map is moved manually
  *
